updated stars as well as jump?
asdf
