--- Assets/Exit.cs
@@ -1,18 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Exit : MonoBehaviour
-{
-    public void exitGame( )
-    {
-        StartCoroutine( goodbye( ) );
-    }
-    
-    public IEnumerator goodbye( )
-    {
-        yield return new WaitForSeconds( 2 );
-        Debug.Log( "exitgame" );
-        Application.Quit( );
-    }
-}
