--- Assets/FadeToBlack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8fb042ce42f472145a623059aba3e2f9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
