+++ Assets/Materials/Stars.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 07e2746b35978cb43b9cf43004cbcecf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
