--- Assets/PlayerMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6cb8f81e448fc1a4899a0f92f0083626
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
