--- Assets/SceneChange.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d9bda92a8524d2f4aa924d84a21bd503
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
