+++ Assets/Scripts/Exit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 252bfaa6e70a4974fa7a656bbcb67b37
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
