+++ Assets/Scripts/FadeToBlack.cs
@@ -1,0 +1,27 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FadeToBlack : MonoBehaviour
+{ 
+    public void FadeMeOut( )
+    {
+        StartCoroutine ( FadeOut( ) );
+    }
+
+    /* Fades scene to black by decrementing alpha over time. */
+    IEnumerator FadeOut( )
+    {
+        CanvasGroup canvasGroup = GetComponent< CanvasGroup >( );
+
+        while( canvasGroup.alpha > 0 )
+        {
+            canvasGroup.alpha -= Time.deltaTime / 2;
+            yield return null;
+        }
+
+        /* This makes sure buttons aren't interactable while fading out. */
+        canvasGroup.interactable = false;
+        yield return null;
+    }
+}
