--- Assets/cloudScroll.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6035d6d670817444ca9add80ce3fb4b7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
