--- Assets/stayPut.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a3aec3135024cf549bcdd5e2da872c73
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
