--- Assets/whenEnterShowThis.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c65cd5b90879bc74d812b5be9e34b7ea
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
