+++ UserSettings/EditorUserSettings.asset
@@ -21,13 +21,13 @@ EditorUserSettings:
       value: 5a04515053065e0c0e5f587446700f4446154e7b2f29766929711b65bbe3353c
       flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 020356575c0d5b5f545f5d21447a0e44474f40287a7b703479711f6be0e1326f
+      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 0705525303000b0b0b5b542312740744124e1b287a2a71627c2f1832b5b63639
+      value: 020356575c0d5b5f545f5d21447a0e44474f40287a7b703479711f6be0e1326f
       flags: 0
     RecentlyUsedSceneGuid-7:
-      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
+      value: 0705525303000b0b0b5b542312740744124e1b287a2a71627c2f1832b5b63639
       flags: 0
     RecentlyUsedSceneGuid-8:
       value: 5408000551560b5e5c5d082648255c44124f4f2e7c717436797b1b30e0e2363a
